+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Changes Since Last Release
 
+### New Features & Functionality
+
+- The `registry login` and `registry logout` commands now support a `--authfile
+  <path>` flag, which causes the OCI credentials to be written to / removed from
+  a custom file located at `<path>` instead of the default location
+  (`$HOME/.singularity/docker-config.json`). The commands `pull`, `push`, `run`,
+  `exec`, and `shell` can now also be passed a `--authfile <path>` option, to
+  read OCI registry credentials from this custom file.
+
 ### Bug Fixes
 
 - Don't bind `/var/tmp` on top of `/tmp` in the container, where `/var/tmp`